requirements.txt
langchain
ipykernel
langchain_community
langchain-text-splitters
langchain_huggingface
langchain_chroma
python-dotenv
sentence_transformers
chromadb
langchain_groq
streamlit
pypdf
fastapi
uvicorn
langserve
python-multipart
aiofiles
